--- view/src/pages/MyPage/consts/mockData.ts
@@ -1,55 +1,0 @@
-export default {
-  assets: [
-    {
-      currencyCode: 'KRW',
-      name: '원화',
-      amount: -1,
-    },
-    {
-      currencyCode: 'BTC',
-      name: '비트코인',
-      amount: -1,
-    },
-    /*
-    {
-      currency_code: 'ETH',
-      name: '이더리움',
-      amount: 125,
-    },
-    {
-      currency_code: 'LTC',
-      name: '라이트코인',
-      amount: 50,
-    },
-    {
-      currency_code: 'XRP',
-      name: '리플',
-      amount: 200,
-    },
-    {
-      currency_code: 'ADA',
-      name: '카르다노',
-      amount: 350,
-    },
-    {
-      currency_code: 'DOT',
-      name: '폴카닷',
-      amount: 75,
-    },
-    {
-      currency_code: 'BCH',
-      name: '비트코인 캐시',
-      amount: 30,
-    },
-    {
-      currency_code: 'SOL',
-      name: '솔라나',
-      amount: 20,
-    },
-    {
-      currency_code: 'LINK',
-      name: '체인링크',
-      amount: 90,
-    },*/
-  ],
-};
