+++ view/src/widgets/CashTransaction/api/getAssetsApi.ts
